Add files via upload
I had written a PowerShell script that displays the statuses (active or inactive) of Active Directory users and also lists which permissions each user has to access which folders. I wrote this to use on the file server side.

The results are extracted in the folder you are in as FolderAccessList.csv.